Bug 36031 - Add a css property "nominal" to bypass number transformation. r=shelly
(cherry picked from commit ba343eace8c17da11f2c87e1f623817bb9a3c4a6)

+++ gfx/src/nsFont.h
@@ -53,7 +53,7 @@ struct nsFont {
 
   // Variant subproperties
   uint8_t variantCaps;
-  uint8_t variantNumeric;
+  uint16_t variantNumeric;
   uint8_t variantPosition;
   uint8_t variantWidth;
